Introduce MANDOC_DB_DIR
This allow the separation of mandoc.db from man paths. Note that this
causes all database related operation to be done in MANDOC_DB_DIR.

OSes such as Haiku use have a package management system that
mark certain folders as read-only (in this case, man directory). This
allows mandocdb to work in such situation.